Fixed the "bad bfOffBits" error
The problem was that the library was not calculating the offset to the
pixel data correctly. There are special cases in the BMP standard that
are a little weird.

Also the BPP in the ICO image entries LIE. Not every ICO generator setx
the fields correctly. So...

NOTE TO READER:
Info in BMP > Info in ICO