Fixed out-of-bounds error in tests.
So the problem was that some of the ICOs like Cloudflares cotained PNGs.
Microsoft recommends that any 256x256 ICO be contained as PNG instead of BMP.
So these ICOs contained PNGs but we were not decoding them properly and
the ICO entry for the PNGs were nearly worthless in terms of
information.
Only Size and Offset were of any use but that is all we need to grab the
PNG data for decoding.