Fixed alpha mask problems.
Worked with @mveytsman to fix the alpha issues, specifically with the
apple.com.ico
There were two issues:

1. One of the either the alpha mask or the image was backwards and
needed to be flipped.

2. We had an off-by-one error in our for loop which caused the alpha
mask to not be applied perfectly.

Also removed error checking when reading the image data into an array.
The problem is that some ICOs lie about the size of the image but the
data is still valid.