Switched tests from PNG to JPG.
To make the tests run.

This change was done because of long-standing issues with the Go PNG
library and third-party PNG libraries:
https://code.google.com/p/go/issues/detail?id=4278
Certain ICOs contain PNGs that will cause the Go PNG lib to spit out a
"chunk out of order" error.
It seems that prior versions of libpng which is what ImageMagick depends
on did not check the chunk ordering of certain chunks.
This means that there are invalid-but-working PNGs floating out there
which this library will fail to decode.

Either a hack needs to be added to this library or to the Go PNG lib in
order to fix this.
Or all remnants of the invalid PNGs need to be purged from the internet,
this is likely impossible.